+++ src/types/point.rs
@@ -69,6 +69,7 @@ where
 mod tests {
     use super::{Coord, Point};
     use crate::{Geometry, Wkt};
+    use std::str::FromStr;
 
     #[test]
     fn basic_point() {